meditation styles added to seperate file
